Add Div, Sub, Expand, and Tile operations (#33)
* add Div and Sub

* add Div and Sub to tf converter

* add expand and tile ops

* begin refactor of simplifiers

* refactor simplifiers
