+++ tests/artifacts/properties/localrobustness.py
@@ -11,7 +11,7 @@ N = Network()[INPUT_LAYER:OUTPUT_LAYER]
 x = np.random.normal(
     loc=eval(os.environ.get("SHIFT", "0")),
     scale=eval(os.environ.get("SCALE", "0")),
-    size=N.input_shape[0],
+    size=(1,) + N.input_shape[0][1:],
 ).astype(np.float32)
 y = np.argmax(N(x))
 epsilon = eval(os.environ.get("EPSILON", "0.01"))
